HW4

+++ common/models/user.go
@@ -8,7 +8,7 @@ import (
 type User struct {
 	// ID пользователя
 	// example: 1
-	ID uint `json:"id" gorm:"primarykey"`
+	ID uint64 `json:"id" gorm:"primarykey"`
 
 	// Логин пользователя
 	// example: johndoe
